Moved Kenzos tests into tests package
